backend/requirements.txt
# Core
fastapi>=0.111.0
uvicorn>=0.30.0
mistralai>=1.2.1
python-dotenv>=1.0.0
supabase>=2.4.0
boto3>=1.35.0

# Google APIs
google-api-python-client>=2.137.0
google-auth>=2.33.0
google-auth-oauthlib>=1.2.0

# Utilities
httpx>=0.27.0
requests>=2.32.0

# PDF Processing
pdf2image>=1.17.0
Pillow>=10.0.0
pdfminer.six>=20231228
PyMuPDF>=1.24.0
python-dotenv>=1.0.1
pdfplumber>=0.11.0
pytesseract>=0.3.10